Little Changes.
The problem is how to bind type parameter as a type
kind. We store `let f : 'a -> 'a = fun x -> [|x|]` in .cmx file, but cmi
includes type information `f : 'b -> 'b`. When outputing this, somehow
compiler forget original type variable ids. But terms includes 'a inside
of the bodies.