Under construction. - create printers (subst, path) - comment out exn handlers (main.ml) - reconsider the position of initialization of Inner_map tables.
Pervasives are loaded before entering Typemod. So we cannnot erase the
info in the entrace of the module.
But if we pass multiple files as command line args,
ocamlopt compile each files in one execution (of course !). So somehow
we need to initialize without ignoring the map of Pervasives.

Accroding to Env.t, initial_env is initialized in Optmain, not Typemod.
By using the env for each compilation, the compoler avoid multiple
unneccesary loading of Pervasives.
I need to do the same thing for Inner_map.map_tbl.

Env.t is immutable, but my tables are mutable.
Fucking hell.